Create characters folder, move CharacterGrid.js into new folder
